Refactor to split the OCI pull logic into a separate package.
